Push Jekyll + Knitr + htmlwidgets helpers
Looks like a lot of docs were also re-rendered with the new Roxygen too,
so that is all being pushed. There are actually a lot of NOTEs in R CMD
Check too, but this is a very lazy commit and doesn't bother to fix any
of it.

Sorry, future self.