Move RPushbullet out of depends
To make things still work with the package moved to Import (meaning it
doesn't automatically attach when we load this package), this commit
adds a function that simulates the important options setting that
usually happens when RPushbullet is loaded interactively.

The big issue here was figuring out how to write to the environment
RPushbullet creates. This is solved here with an assign() where envir =
RPushbullet:::.pkgenv.

Whew...